(PE-15457) Add script for adding git refs to POs
Add the script `src/leiningen/i18n/bin/add-gitref.sh`, which takes a PO
file as its only argument and adds or replaces a PO header entry for
`X-Git-Ref` with the current git HEAD's SHA as the value.

The script is tested in the `puppeetlabs.i18n.bin-test` test namespace.