(PE-15457) Add `update-pot.sh` workflow script.
The `update-pot.sh` script runs the full workflow for updating the
`locales/messages.pot` for clojure projects. The script:
  * moves the current POT file to a temporary location;
  * generates a new POT file;
  * compares the two POT files to determine whether any strings were
    added or removed;
  * if changes were found, commits the new POT file after removing any
    line numbers and adding the current git ref, and pushes the commit
    to the git origin on the branch given by the script's only argument;
  * if no changes were found, restores the extant POT file.

Since this script will be responsible for updating the POT file in a CI
job, change the `i18n` target in the Makefile to not update the POT so
normal developer operations like starting the REPL or running tests
don't update the POT file as well.